+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.12.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.11.6...@coveo/atomic@1.12.0) (2021-11-11)
+
+
+### Features
+
+* **atomic:** add support for focus-visible ([#1463](https://github.com/coveo/ui-kit/issues/1463)) ([7601409](https://github.com/coveo/ui-kit/commit/7601409bdde34269bc34fd037be4d61f798d8e5f))
+
+
+
+
+
+## [1.11.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.11.5...@coveo/atomic@1.11.6) (2021-11-11)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [1.11.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.11.4...@coveo/atomic@1.11.5) (2021-11-10)
+
+
+### Bug Fixes
+
+* **atomic:** fix storybook build for windows user ([#1460](https://github.com/coveo/ui-kit/issues/1460)) ([48d11f0](https://github.com/coveo/ui-kit/commit/48d11f0f1454c6c747548628f1fa6b2e17cb673e))
+
+
+
+
+
+## [1.11.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.11.3...@coveo/atomic@1.11.4) (2021-11-10)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
 ## [1.11.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.11.2...@coveo/atomic@1.11.3) (2021-11-10)
 
 **Note:** Version bump only for package @coveo/atomic